+++ src/resolvers.ts
@@ -1,86 +1,8 @@
 import { existsSync } from 'node:fs'
 import { defu } from 'defu'
-import { join, relative, resolve } from 'pathe'
-import { findPath, useNuxt, tryResolveModule, resolveAlias, resolvePath } from '@nuxt/kit'
+import { join, relative } from 'pathe'
+import { useNuxt, tryResolveModule, resolvePath } from '@nuxt/kit'
 import type { EditorSupportConfig, ExposeConfig, InjectPosition, ModuleOptions, ViewerConfig } from './types'
-
-/**
- * Resolves all configPath values for an application
- *
- * @param path configPath for a layer
- * @returns array of resolved paths
- */
-const resolveConfigPath = async (path: ModuleOptions['configPath'], nuxtOptions = useNuxt().options) =>
-  Promise.all(
-    (Array.isArray(path) ? path : [path])
-      .filter(Boolean)
-      .map(path => path.startsWith(nuxtOptions.buildDir) ? path : findPath(path, { extensions: ['.js', '.cjs', '.mjs', '.ts'] })),
-  ).then(paths => paths.filter((p): p is string => Boolean(p)))
-
-/**
- *
- * @param srcDir
- * @returns array of resolved content globs
- */
-const resolveContentPaths = (srcDir: string, nuxtOptions = useNuxt().options) => {
-  const r = (p: string) => p.startsWith(srcDir) ? p : resolve(srcDir, p)
-  const extensionFormat = (s: string[]) => s.length > 1 ? `.{${s.join(',')}}` : `.${s.join('') || 'vue'}`
-
-  const defaultExtensions = extensionFormat(['js', 'ts', 'mjs'])
-  const sfcExtensions = extensionFormat(Array.from(new Set(['.vue', ...nuxtOptions.extensions])).map(e => e.replace(/^\.*/, '')))
-
-  const importDirs = [...(nuxtOptions.imports?.dirs || [])].map(r)
-  const [composablesDir, utilsDir] = [resolve(srcDir, 'composables'), resolve(srcDir, 'utils')]
-
-  if (!importDirs.includes(composablesDir)) importDirs.push(composablesDir)
-  if (!importDirs.includes(utilsDir)) importDirs.push(utilsDir)
-
-  return [
-    r(`components/**/*${sfcExtensions}`),
-    ...(() => {
-      if (nuxtOptions.components) {
-        return (Array.isArray(nuxtOptions.components) ? nuxtOptions.components : typeof nuxtOptions.components === 'boolean' ? ['components'] : nuxtOptions.components.dirs).map(d => `${resolveAlias(typeof d === 'string' ? d : d.path)}/**/*${sfcExtensions}`)
-      }
-      return []
-    })(),
-
-    nuxtOptions.dir.layouts && r(`${nuxtOptions.dir.layouts}/**/*${sfcExtensions}`),
-    ...([true, undefined].includes(nuxtOptions.pages) ? [r(`${nuxtOptions.dir.pages}/**/*${sfcExtensions}`)] : []),
-
-    nuxtOptions.dir.plugins && r(`${nuxtOptions.dir.plugins}/**/*${defaultExtensions}`),
-    ...importDirs.map(d => `${d}/**/*${defaultExtensions}`),
-
-    r(`{A,a}pp${sfcExtensions}`),
-    r(`{E,e}rror${sfcExtensions}`),
-    r(`app.config${defaultExtensions}`),
-    !nuxtOptions.ssr && nuxtOptions.spaLoadingTemplate !== false && r(typeof nuxtOptions.spaLoadingTemplate === 'string' ? nuxtOptions.spaLoadingTemplate : 'app/spa-loading-template.html'),
-  ].filter((p): p is string => Boolean(p))
-}
-
-/**
- *
- * @param configPath
- * @param nuxt
- * @returns [configuration paths, default resolved content paths]
- */
-export const resolveModulePaths = async (configPath: ModuleOptions['configPath'], nuxt = useNuxt()) => {
-  const mainPaths: [string[], string[]] = [await resolveConfigPath(configPath), resolveContentPaths(nuxt.options.srcDir, nuxt.options)]
-
-  if (Array.isArray(nuxt.options._layers) && nuxt.options._layers.length > 1) {
-    const layerPaths = await Promise.all(
-      nuxt.options._layers.slice(1).map(async (layer): Promise<[string[], string[]]> => ([
-        await resolveConfigPath(layer?.config?.tailwindcss?.configPath || join(layer.cwd, 'tailwind.config'), nuxt.options),
-        resolveContentPaths(layer?.config?.srcDir || layer.cwd, defu(layer.config, nuxt.options) as typeof nuxt.options),
-      ])))
-
-    layerPaths.forEach(([configPaths, contentPaths]) => {
-      mainPaths[0].unshift(...configPaths)
-      mainPaths[1].unshift(...contentPaths)
-    })
-  }
-
-  return mainPaths
-}
 
 /**
  *
